Modifier class
The modifier class manages the setting of parameters for a specific function

Parameters are represented inside the modifier as fields. These fields can have certain types including integer, options, text (not yet implemented)

Fields can have restrictions on their value. e.g. integer fields will have a range of numbers that are valid, options will have a list of options.

On the gui side these fields are converted into components which can display them, and check whether they are valid. The user can interact with the field component as they'd like, and after they are finished they can click "set" which checks if all fields are valid. If they are the component allows the changes