click method
iGP is dumb and sometimes makes a magical curtain appear on the screen for a short amount of time after loading a page. This click method seeks to counteract the magical curtain by trying to click repeatedly. 10 failures and it will through an error is it is likely there is another issue entirely